Fix: Avoid creating IDs for org-refile/org-capture
This commit contains two changes:

(1) Restore the previous behaviour of avoiding ID creation during org-refile and
org-capture calls in the org-bookmark-heading-make-record function, by
updating the detection mechanism in response to upstream org-mode changes.

The previous mechanism (added in 2018) checked for for the bookmark-name variable being
bound to certain values, but bookmark-name is no longer visible to this package,
confirmed by debugging with edebug: (boundp 'bookmark-name) always returns nil.
The org-refile.el file was added upstream in 2020, with lexical binding, as a potential cause.

The new mechanism detects org-refile and org-capture calls using flags, which are toggled before
and after the upstream calls from org-refile and org-capture-store-last-position.

The flags are intentionally set globally, beyond org-mode, as calls like
org-capture are initiated outside of org-mode.

(2) When making bookmark records, always set the handler, even if not setting the ID.

Without this change, these org-bookmark-heading style records with no handler field
cause odd behaviour when jumping to their bookmarks.